参数的调整

+++ Users/delay.c
@@ -1,8 +1,8 @@
 #include "delay.h"
 #include"tim.h"
-void delay(uint32_t us)
+void delay(uint16_t us)
 {
-    uint32_t counter=0;
+    uint16_t counter=0;
     __HAL_TIM_SetCounter(&htim3,counter);
     HAL_TIM_Base_Start(&htim3);
